requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
requests
redis
celery
python-dotenv
PyGithub
python-gitlab
Pillow
chardet
pygments
gitpython
psycopg2-binary
gunicorn
bleach
markdown
python-dateutil
Babel
stripe
Flask-WTF
email-validator
SQLAlchemy
flask_session
